New NYC Weather report on line 10
Still having some confusion with some aspects of this.  I created a pull request a few days ago but never merged.  Now it's got a lot of obsolete info in it and does not include a lot of other students' edits.  I just want to ditch the request and re-pull the whole github playground to have a fresh copy but it doesn't seem that simple, and i don't want to blow up everyone else's work. Maybe after the week is done with I will...